// frontend/src/app/shared/api/model/list-users200-response-results-inner.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ListUsers200ResponseResultsInnerActivated } from './list-users200-response-results-inner-activated';
import { UserRole } from './user-role';


export interface ListUsers200ResponseResultsInner { 
  [key: string]: any | any;


    /**
     * Unikalny identyfikator
     */
    idUser: number;
    name: string;
    surname: string;
    email: string;
    role: UserRole;
    active: boolean;
    activated: ListUsers200ResponseResultsInnerActivated;
    created: string;
}
export namespace ListUsers200ResponseResultsInner {
}